Renamed to hexdata and added package-info.java
